Separate kubernetes tests into separate cloud providers (#853)
* Move Azure specific account and client classes from Azure Sashimi to Calamari.CloudAccounts

* Implement Kubernetes Discovery Command for AKS (Azure)

* Add TargetTags extension method

* Add type property to AccountAuthenticationDetails so that it can be set in tests.

* Make WriteServiceMessage method virtual so that tests can intercept serviceMessages

* Add relevant discovery tags to aks cluster

* Make Kubernetes base fixture inherit from CalamariFixture

* Add Integration test for AKS cluster discovery

* Move Azure specific discovery code to new Calamari.Azure project

so that Calamari project can still be optionally build against .net 4.0

* Remove nullables as not compatible with c# version under 8.0

* Update to Enumerable of KeyValuePairs as IReadOnlyDictionary doesn't exist in .Net 4.0

* Fix build issue with Azure Web App Helper

* Remove nullable ? from AzureWebAppHelper

* Fix issues with new test

* Fix issue with potentially discovering other matching clusters from other parallel tests

* Add Type property to ITargetDiscoveryAuthenticationDetails to match Server version

* Revert adding Azure Client classes to CloudAccounts

In the end I don't think it's worth the trouble as we only use a small subset of these classes

* Add required AzureClient related classes to Calamari.Azure

* Removing Jetbrains annotations from Calamari.Azure.csproj

* Remove jetbrains annotations from Calamari project

* Remove trying to get health check container from existing deployment variables

as they aren't set yet. We'll add this back in later.

* Rename method from DiscoveryClusters to DiscoverClusters

* Fix weird formatting in AzureKubernetesDiscovery

* Update DiscoveryCommand to return non-zero code if parsing json fails.

* Fix name of AzureKubernetesDiscoverer

* Add AwsKubernetesDiscoverer and AwsAuthenticationDetails classes

* Move special logger to only be used in new test

* Add ExitStatus static class with appropriate values for different Calamari errors

* Remove all health check container implementation code to be fully implemented under that ticket.

* Fix authentication issue for AWS account credentials

* Implement return of discovered clusters for AWS credentials

* Fix cluster name to be using proper cluster name instead of name

* Add verbose logging to AwsKubernetesDiscoverer.cs

* Update logging

* Remove newer c# feature to increase compatibility

* Adding test for EKS clusters

* Fix extensions for EKS test

(cherry picked from commit ca392c17768877806cc54faeaf57bfbb9025c28b)

* Update region log to use TestContext.Progress

(cherry picked from commit 5650de657bf5984a778daa5f667617cf08980f09)

* Add tests and fix resulting bugs

(cherry picked from commit 9927e311f81fb793ff077ec403d13851d85064b5)

# Conflicts:
#	source/Calamari.Tests/KubernetesFixtures/KubernetesContextScriptWrapperLiveFixture.cs

* fix ci build error

* Add Role Session Name and Session Duration to test

* Fix tests

* Fix session duration to valid value

* fix incorrect awsAssumeRoleSessionDurationSeconds property name

* Address code review comments

* Address code review comments #2

* Fix wording of warning

* Update AwsAuthentication to use InstanceProfileAwsCredentials if no account is specified.

* Add tests to ensure EC2 Instance authentication works for Kube Target Discovery

* Fix logging and error handling

* separate test executions to run them sequentially

* Move shared code to base class and fix logging

* Initial separation

* Fix using statements

* Remove unused usings

* Add test for when no valid credentials are given

* Move test to correct location after merge

* Fix no valid credentials test

* Fix failure on azure tests

* Fix exception handling of credentials

* Update to only use single kubeconfig.tpl as each cloud provider uses the same config.

* Update to not use the ".." method for going up a folder in a path as might be causing issues on debian

* Push logging to help diagnose test failure

* Update to use TestProgress instead of Console...

* Fix issue with kubeconfig.tpl
