+++ source/Calamari.Tests/KubernetesFixtures/Terraform/Clusters/AKS/aks.kubernetes.tf
