+++ source/Calamari.Tests/KubernetesFixtures/Terraform/Clusters/AKS/aks.tf
