+++ source/Calamari.Tests/KubernetesFixtures/Terraform/Clusters/EKS/eks.tf
