+++ source/Calamari.Tests/KubernetesFixtures/Terraform/Clusters/EKS/main.tf
@@ -1,0 +1,5 @@
+variable "tests_source_dir" {
+  type = string
+}
+
+resource "random_pet" "prefix" {}
