+++ source/Calamari.Tests/KubernetesFixtures/Terraform/Clusters/EKS/outputs.tf
@@ -1,0 +1,44 @@
+output "eks_client_id" {
+  value = aws_iam_access_key.default.id
+}
+
+output "eks_secret_key" {
+  value     = aws_iam_access_key.default.secret
+  sensitive = true
+}
+
+output "eks_iam_role_arn" {
+  value = aws_iam_role.user.arn
+}
+
+output "eks_cluster_endpoint" {
+  description = "Endpoint for EKS control plane."
+  value       = aws_eks_cluster.default.endpoint
+}
+
+output "eks_cluster_ca_certificate" {
+  value     = base64decode(aws_eks_cluster.default.certificate_authority[0].data)
+  sensitive = true
+}
+
+output "eks_cluster_name" {
+  description = "EKS name."
+  value       = aws_eks_cluster.default.name
+}
+
+output "eks_cluster_arn" {
+  description = "EKS ARN"
+  value       = aws_eks_cluster.default.arn
+}
+
+output "aws_vpc_id" {
+  value = aws_vpc.default.id
+}
+
+output "aws_subnet_id" {
+  value = aws_subnet.default[0].id
+}
+
+output "aws_iam_instance_profile_name" {
+  value = aws_iam_instance_profile.profile.name
+}
