+++ source/Calamari.Tests/KubernetesFixtures/Terraform/Clusters/EKS/providers.tf
@@ -1,0 +1,16 @@
+terraform {
+  required_providers {
+    aws = {
+      source  = "hashicorp/aws"
+      version = ">= 3.45.0"
+    }
+    kubernetes = {
+      source  = "hashicorp/kubernetes"
+      version = ">= 2.3.2"
+    }
+  }
+  required_version = ">= 0.15"
+}
+
+provider "aws" {
+}
