+++ source/Calamari.Tests/KubernetesFixtures/Terraform/Clusters/GKE/gke.tf
