+++ source/Calamari.Tests/KubernetesFixtures/Terraform/Clusters/GKE/main.tf
@@ -1,0 +1,1 @@
+resource "random_pet" "prefix" {}
