+++ source/Calamari.Tests/KubernetesFixtures/Terraform/Clusters/GKE/providers.tf
@@ -1,13 +1,5 @@
 terraform {
   required_providers {
-    azurerm = {
-      source  = "hashicorp/azurerm"
-      version = "= 2.78.0" #Due to a bug with 2.79.x we are pinning to this version for now. https://github.com/Azure/AKS/issues/2584 
-    }
-    aws = {
-      source  = "hashicorp/aws"
-      version = ">= 3.45.0"
-    }
     google = {
       source  = "hashicorp/google"
       version = "3.71.0"
@@ -18,13 +10,6 @@ terraform {
     }
   }
   required_version = ">= 0.15"
-}
-
-provider "azurerm" {
-  features {}
-}
-
-provider "aws" {
 }
 
 provider "google" {
